# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pyKorf",
    "version": "0.18.0"
  },
  "paths": {
...
    "/api/doc-register/search-eddr": {
      "get": {
        "tags": [
          "doc-register"
        ],
        "summary": "Api Search Eddr",
        "description": "Search EDDR entries by title.",
        "operationId": "searchEddr",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocRegisterSearchEddrResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/doc-register/search-query": {
      "get": {
        "tags": [
          "doc-register"
        ],
        "summary": "Api Search Query",
        "description": "Search query entries by document number.",
        "operationId": "searchQuery",
        "parameters": [
          {
            "name": "doc_no",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Doc No"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocRegisterSearchQueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/doc-register/search-files": {
      "get": {
        "tags": [
          "doc-register"
        ],
        "summary": "Api Search Files",
        "description": "Search query entries by name or path.",
        "operationId": "searchFiles",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocRegisterSearchFilesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/doc-register/rebuild-db": {
      "post": {
        "tags": [
          "doc-register"
        ],
        "summary": "Api Rebuild Db",
        "description": "Force rebuild the Document Register database from Excel.",
        "operationId": "rebuildDocRegisterDb",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocRegisterRebuildResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/doc-register/config": {
      "post": {
        "tags": [
          "doc-register"
        ],
        "summary": "Api Config",
        "description": "Save Document Register configuration.",
        "operationId": "setDocRegisterConfig",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetDocRegisterConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocRegisterConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DocRegisterConfigResponse": {
        "properties": {
          "excel_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excel Path"
          },
          "sp_site_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sp Site Url"
          }
        },
        "type": "object",
        "title": "DocRegisterConfigResponse"
      },
      "DocRegisterRebuildResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": false
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "stats": {
            "additionalProperties": true,
            "type": "object",
            "title": "Stats",
            "default": {}
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "title": "DocRegisterRebuildResponse"
      },
      "DocRegisterSearchEddrResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/EddrResult"
            },
            "type": "array",
            "title": "Results",
            "default": []
          }
        },
        "type": "object",
        "title": "DocRegisterSearchEddrResponse"
      },
      "DocRegisterSearchFilesResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/QueryEntryResult"
            },
            "type": "array",
            "title": "Results",
            "default": []
          }
        },
        "type": "object",
        "title": "DocRegisterSearchFilesResponse"
      },
      "DocRegisterSearchQueryResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/QueryEntryResult"
            },
            "type": "array",
            "title": "Results",
            "default": []
          }
        },
        "type": "object",
        "title": "DocRegisterSearchQueryResponse"
      },
...
      "EddrResult": {
        "properties": {
          "document_no": {
            "type": "string",
            "title": "Document No"
          },
          "title": {
            "type": "string",
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "document_no",
          "title"
        ],
        "title": "EddrResult"
      },
...
      "EmptyRequest": {
        "properties": {},
        "type": "object",
        "title": "EmptyRequest",
        "description": "Explicit empty request payload for body-less POST endpoints."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QueryEntryResult": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "modified": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified"
          },
          "modified_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified By"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "item_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Type"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "QueryEntryResult"
      },
...
      "SetDocRegisterConfigRequest": {
        "properties": {
          "excel_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excel Path"
          },
          "sp_site_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sp Site Url"
          }
        },
        "type": "object",
        "title": "SetDocRegisterConfigRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}